Added Spring Security on user and journal entries

+++ src/main/java/com/hamdan/journalapp/service/JournalEntityService.java
@@ -52,13 +52,23 @@ public class JournalEntityService
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName)
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
-
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed)
+            {
+                userService.saveEntry(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e)
+        {
+                throw new RuntimeException("An error occured while deleting the entry.", e);
+        }
+        return removed;
     }
 
 }
